refactor(tasks): move G1 core tracking owner [skip ci]
