--- src/unilab/envs/motion_tracking/g1/__init__.py
@@ -1,23 +1,0 @@
-"""Motion tracking environments for Unitree G1."""
-
-from .tracking import (
-    G1MotionTracking23DofCfg,
-    G1MotionTracking23DofDeployEnvCfg,
-    G1MotionTracking23DofEnvCfg,
-    G1MotionTrackingCfg,
-    G1MotionTrackingDeployEnv,
-    G1MotionTrackingDeployEnvCfg,
-    G1MotionTrackingEnv,
-    G1MotionTrackingEnvCfg,
-)
-
-__all__ = [
-    "G1MotionTracking23DofCfg",
-    "G1MotionTracking23DofDeployEnvCfg",
-    "G1MotionTracking23DofEnvCfg",
-    "G1MotionTrackingCfg",
-    "G1MotionTrackingDeployEnv",
-    "G1MotionTrackingDeployEnvCfg",
-    "G1MotionTrackingEnv",
-    "G1MotionTrackingEnvCfg",
-]
