+++ src/unilab/tasks/__init__.py
@@ -17,7 +17,6 @@ __unilab_registry_modules__ = (
     "unilab.tasks.manipulation.allegro_inhand",
     "unilab.tasks.manipulation.sharpa_inhand",
     "unilab.tasks.manipulation.stewart",
-    "unilab.envs.motion_tracking.g1",
     "unilab.tasks.motion_tracking.g1",
     "unilab.tasks.motion_tracking.x2",
 )
