+++ src/unilab/tasks/motion_tracking/__init__.py
@@ -20,6 +20,12 @@ from .g1 import (
     G1FlipTrackingSAC23DofEnv,
     G1FlipTrackingSACCfg,
     G1FlipTrackingSACEnv,
+    G1MotionTracking23DofCfg,
+    G1MotionTracking23DofDeployEnvCfg,
+    G1MotionTracking23DofEnvCfg,
+    G1MotionTrackingCfg,
+    G1MotionTrackingEnv,
+    G1MotionTrackingEnvCfg,
     G1MotionTrackingSAC23DofCfg,
     G1MotionTrackingSAC23DofEnv,
     G1MotionTrackingSACCfg,
@@ -64,6 +70,12 @@ __all__ = [
     "G1FlipTrackingSAC23DofEnv",
     "G1FlipTrackingSACCfg",
     "G1FlipTrackingSACEnv",
+    "G1MotionTracking23DofCfg",
+    "G1MotionTracking23DofDeployEnvCfg",
+    "G1MotionTracking23DofEnvCfg",
+    "G1MotionTrackingCfg",
+    "G1MotionTrackingEnv",
+    "G1MotionTrackingEnvCfg",
     "G1MotionTrackingSAC23DofCfg",
     "G1MotionTrackingSAC23DofEnv",
     "G1MotionTrackingSACCfg",
