+++ src/unilab/tasks/motion_tracking/g1/flip_tracking.py
@@ -16,7 +16,8 @@ from unilab.envs.motion_tracking.common.config import (
     _zero_pose_randomization,
     _zero_velocity_randomization,
 )
-from unilab.envs.motion_tracking.g1.tracking import (
+
+from .tracking import (
     G1MotionTrackingCfg,
     G1MotionTrackingEnv,
     PoseRandomization,
