+++ tests/tasks/test_package_boundary.py
@@ -21,7 +21,6 @@ _TASK_REGISTRY_MODULES = (
     "unilab.tasks.manipulation.allegro_inhand",
     "unilab.tasks.manipulation.sharpa_inhand",
     "unilab.tasks.manipulation.stewart",
-    "unilab.envs.motion_tracking.g1",
     "unilab.tasks.motion_tracking.g1",
     "unilab.tasks.motion_tracking.x2",
 )
